requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
gunicorn==21.2.0
